+++ app/src/main/java/com/example/limjinlok/ContactListAdapter.kt
@@ -1,7 +1,11 @@
 package com.example.limjinlok
 
 import android.view.LayoutInflater
+import android.view.View
 import android.view.ViewGroup
+import android.widget.ImageButton
+import android.widget.ImageView
+import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.example.limjinlok.model.ContactListData
 
@@ -9,64 +13,52 @@ class ContactListAdapter(
     val mItems: ArrayList<ContactListData>,
     val actions: (Int, ContactListData) -> Unit
 ) :
-    RecyclerView.Adapter<ContactViewHolder>() {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ContactViewHolder {
-        val view =
-            LayoutInflater.from(parent.context).inflate(R.layout.item_contactlist, parent, false)
-        return ContactViewHolder(view, actions)
-class ContactListAdapter(val mItems: ArrayList<ContactListData>) :
     RecyclerView.Adapter<RecyclerView.ViewHolder>() {
     private val TYPE_EVEN = 0
     private val TYPE_ODD = 1
     private val TYPE_GRID = 2
 
     override fun getItemViewType(position: Int): Int {
-        return when(position % 2) {
+        return when (position % 2) {
             0 -> TYPE_EVEN
             1 -> TYPE_ODD
             else -> TYPE_GRID
         }
-class ContactListAdapter(
-    val mItems: ArrayList<ContactListData>,
-    val actions: (Int, ContactListData) -> Unit
-) :
-    RecyclerView.Adapter<ContactViewHolder>() {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ContactViewHolder {
-        val view =
-            LayoutInflater.from(parent.context).inflate(R.layout.item_contactlist, parent, false)
-        return ContactViewHolder(view, actions)
-
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         return when (viewType) {
             TYPE_EVEN -> {
-                val view = LayoutInflater.from(parent.context).inflate(R.layout.item_contactlist, parent, false)
-                ContactViewHolder(view)
+                val view = LayoutInflater.from(parent.context)
+                    .inflate(R.layout.item_contactlist, parent, false)
+                ContactViewHolder(view, actions)
             }
+
             TYPE_ODD -> {
-                val view = LayoutInflater.from(parent.context).inflate(R.layout.item_contactlist_odd, parent, false)
-                ContactViewHolderOdd(view)
+                val view = LayoutInflater.from(parent.context)
+                    .inflate(R.layout.item_contactlist_odd, parent, false)
+                ContactViewHolderOdd(view, actions)
             }
+
             TYPE_GRID -> {
-                val view = LayoutInflater.from(parent.context).inflate(R.layout.item_contactlist_grid, parent, false)
-                ContactViewHolderGrid(view)
+                val view = LayoutInflater.from(parent.context)
+                    .inflate(R.layout.item_contactlist_grid, parent, false)
+                ContactViewHolderGrid(view, actions)
             }
+
             else -> throw IllegalArgumentException("Invalid view type")
         }
     }
 
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         val item: ContactListData = mItems[position]
-        when (holder){
+        when (holder) {
             is ContactViewHolder -> {
                 holder.profileImage.setImageResource(item.userImage)
                 holder.tv_name.text = item.userData[0].content
                 holder.tv_nickname.text = item.userData[1].content
                 holder.favBut.setOnClickListener {
-                    item.isFavorite =!item.isFavorite
+                    item.isFavorite = !item.isFavorite
                     if (item.isFavorite)
                         holder.favBut.setImageResource(R.drawable.staron)
                     else
@@ -97,12 +89,13 @@ class ContactListAdapter(
                     holder.tv_name.text = item.userData[0].content
                 }
             }
+
             is ContactViewHolderOdd -> {
                 holder.Od_profileImage.setImageResource(item.userImage)
                 holder.Od_tv_name.text = item.userData[0].content
                 holder.Od_tv_nickname.text = item.userData[1].content
                 holder.Od_favBut.setOnClickListener {
-                    item.isFavorite =!item.isFavorite
+                    item.isFavorite = !item.isFavorite
                     if (item.isFavorite)
                         holder.Od_favBut.setImageResource(R.drawable.staron)
                     else
@@ -133,6 +126,7 @@ class ContactListAdapter(
                     holder.Od_tv_name.text = item.userData[0].content
                 }
             }
+
             is ContactViewHolderGrid -> {
                 holder.gr_profileImage.setImageResource(item.userImage)
                 holder.gr_tv_name.text = item.userData[0].content
@@ -162,15 +156,6 @@ class ContactListAdapter(
                     holder.gr_tv_name.text = item.userData[0].content
                 }
             }
-        holder.profileImage.setImageResource(item.userImage)
-        holder.tv_name.text = item.userData[0].content
-        holder.tv_nickname.text = item.userData[1].content
-        if (mItems[position].isFavorite)
-            holder.favBut.setImageResource(R.drawable.staron)
-        else
-            holder.favBut.setImageResource(R.drawable.staroff)
-        holder.itemView.setOnClickListener {
-            actions(position, item)
         }
     }
 
@@ -179,6 +164,28 @@ class ContactListAdapter(
         return mItems.size
     }
 
+    class ContactViewHolder(itemView: View, val actions: (Int, ContactListData) -> Unit) :
+        RecyclerView.ViewHolder(itemView) {
+        val profileImage: ImageView = itemView.findViewById(R.id.ProfileImg)
+        val tv_name: TextView = itemView.findViewById(R.id.tv_name)
+        val tv_nickname: TextView = itemView.findViewById(R.id.tv_nickname)
+        val favBut: ImageButton = itemView.findViewById(R.id.btn_fav)
+    }
+
+    class ContactViewHolderGrid(itemView: View, val actions: (Int, ContactListData) -> Unit) :
+        RecyclerView.ViewHolder(itemView) {
+        val gr_profileImage: ImageView = itemView.findViewById(R.id.ProfileImggrid)
+        val gr_tv_name: TextView = itemView.findViewById(R.id.tv_namegrid)
+        val gr_tv_nickname: TextView = itemView.findViewById(R.id.tv_nicknamegrid)
+    }
+
+    class ContactViewHolderOdd(itemView: View, val actions: (Int, ContactListData) -> Unit) :
+        RecyclerView.ViewHolder(itemView) {
+        val Od_profileImage: ImageView = itemView.findViewById(R.id.ProfileImgodd)
+        val Od_tv_name: TextView = itemView.findViewById(R.id.tv_nameodd)
+        val Od_tv_nickname: TextView = itemView.findViewById(R.id.tv_nicknameodd)
+        val Od_favBut: ImageButton = itemView.findViewById(R.id.btn_favodd)
+    }
 
 }
